add purchase in user

+++ src/main/java/com/example/demo/entity/User.java
@@ -25,7 +25,7 @@ public class User {
     private Long suspend_number;
     private Double[] rating;
     private String[] favorite;
-    private Double[] purchase;
+    private Long[] purchase;
 
     private Long[] recent_view;
